+++ marshmallow/__init__.py
@@ -11,7 +11,7 @@ from marshmallow.utils import EXCLUDE, INCLUDE, RAISE, pprint, missing
 from marshmallow.exceptions import ValidationError
 from distutils.version import LooseVersion
 
-__version__ = '3.0.0rc2'
+__version__ = '3.0.0rc4.dev0'
 __version_info__ = tuple(LooseVersion(__version__).version)
 __author__ = 'Steven Loria'
 __all__ = [